summarizer/requirements.txt
openai
requests
beautifulsoup4
PyPDF2
python-docx
langdetect
deep-translator